COC-74 Layout moved and Navigation modified
