+++ app/src/main/java/com/bifrost/cocinarte/models/main/AccountProfileViewModel.kt
@@ -1,4 +1,4 @@
-package com.bifrost.cocinarte.models.login
+package com.bifrost.cocinarte.models.main
 
 import androidx.lifecycle.ViewModel
