+++ app/src/main/res/layout/account_profile_fragment.xml
@@ -5,7 +5,7 @@
     android:id="@+id/accountProfileLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.login.AccountProfileFragment">
+    tools:context=".fragments.main.AccountProfileFragment">
 
     <TextView
         android:id="@+id/txtAccountProfileTitle"
