+++ app/src/main/res/layout/delete_account_fragment.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.login.DeleteAccountFragment">
+    tools:context=".fragments.main.DeleteAccountFragment">
 
     <TextView
         android:id="@+id/txtDeleteAccountTitle"
