+++ app/src/main/res/layout/log_out_fragment.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout3"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.login.LogOutFragment">
+    tools:context=".fragments.main.LogOutFragment">
 
     <TextView
         android:id="@+id/textView3"
