+++ app/src/main/res/navigation/main_navgraph.xml
@@ -27,6 +27,12 @@
         <action
             android:id="@+id/action_userProfileFragment_to_favouritesFragment"
             app:destination="@id/favouritesFragment" />
+        <action
+            android:id="@+id/action_userProfileFragment_to_accountProfileFragment"
+            app:destination="@id/accountProfileFragment" />
+        <action
+            android:id="@+id/action_userProfileFragment_to_logOutFragment"
+            app:destination="@id/logOutFragment" />
     </fragment>
     <activity
         android:id="@+id/preferenceActivity"
@@ -57,4 +63,36 @@
         android:name="com.bifrost.cocinarte.fragments.main.PreparedFragment"
         android:label="prepared_fragment"
         tools:layout="@layout/prepared_fragment" />
+    <fragment
+        android:id="@+id/accountProfileFragment"
+        android:name="com.bifrost.cocinarte.fragments.main.AccountProfileFragment"
+        android:label="account_profile_fragment"
+        tools:layout="@layout/account_profile_fragment" >
+        <action
+            android:id="@+id/action_accountProfileFragment_to_deleteAccountFragment"
+            app:destination="@id/deleteAccountFragment" />
+    </fragment>
+    <fragment
+        android:id="@+id/deleteAccountFragment"
+        android:name="com.bifrost.cocinarte.fragments.main.DeleteAccountFragment"
+        android:label="delete_account_fragment"
+        tools:layout="@layout/delete_account_fragment" >
+        <action
+            android:id="@+id/action_deleteAccountFragment_to_loginActivity"
+            app:destination="@id/loginActivity" />
+    </fragment>
+    <fragment
+        android:id="@+id/logOutFragment"
+        android:name="com.bifrost.cocinarte.fragments.main.LogOutFragment"
+        android:label="log_out_fragment"
+        tools:layout="@layout/log_out_fragment" >
+        <action
+            android:id="@+id/action_logOutFragment_to_loginActivity"
+            app:destination="@id/loginActivity" />
+    </fragment>
+    <activity
+        android:id="@+id/loginActivity"
+        android:name="com.bifrost.cocinarte.activities.LoginActivity"
+        android:label="activity_login"
+        tools:layout="@layout/activity_login" />
 </navigation>
